==> notebooks/02_data_validation.ipynb <==
import os

%pwd

os.chdir("../")

%pwd

import numpy as np
import pandas as pd
import warnings

warnings.filterwarnings("ignore")

df = pd.read_csv("./artifacts/data_ingestion/travel.csv")
df.head()

df.isnull().sum()

df["Gender"].value_counts()

df["MaritalStatus"].value_counts()

df["TypeofContact"].value_counts()

df["Gender"] = df["Gender"].replace("Fe Male", "Female")
df["MaritalStatus"] = df["MaritalStatus"].replace("Single", "Unmarried")

df.head()

df["Gender"].value_counts()

na_features = [features for features in df.columns if df[features].isnull().sum() >= 1]
for feature in na_features:
    print(feature, np.round(df[feature].isnull().mean() * 100, 5), "% missing values")

df.columns

df["Age"].fillna(df["Age"].median(), inplace=True)
df["TypeofContact"].fillna(df["TypeofContact"].mode()[0], inplace=True)
df["DurationOfPitch"].fillna(df["DurationOfPitch"].median(), inplace=True)
df["NumberOfFollowups"].fillna(df["NumberOfFollowups"].mode()[0], inplace=True)
df["PreferredPropertyStar"].fillna(df["PreferredPropertyStar"].mode()[0], inplace=True)
df["NumberOfTrips"].fillna(0, inplace=True)
df["NumberOfChildrenVisiting"].fillna(
    df["NumberOfChildrenVisiting"].mode()[0], inplace=True
)
df["MonthlyIncome"].fillna(df["MonthlyIncome"].median(), inplace=True)

df.isnull().sum()

df.drop("CustomerID", axis=1, inplace=True)
df["TotalVisiting"] = (
    df["NumberOfChildrenVisiting"] + df["NumberOfPersonVisiting"]
)
df.drop(
    ["NumberOfChildrenVisiting", "NumberOfPersonVisiting"], axis=1, inplace=True
)

df.columns

df.info()

df.to_csv("./data/processed/travel_processed.csv", index=False)

from dataclasses import dataclass
from pathlib import Path


@dataclass(frozen=True)
class DataValidationConfig:
    root_dir: Path
    raw_data_path: Path
    processed_data_path: Path
    all_schema: dict
    STATUS_FILE: Path

from src.constants import *
from src.utils.common import read_yaml, create_directories

class ConfigurationManager:
    def __init__(
        self,
        config_filepath=CONFIG_FILE_PATH,
        model_config_filepath=MODEL_CONFIG_FILE_PATH,
        schema_filepath=SCHEMA_FILE_PATH,
    ):

        self.config = read_yaml(config_filepath)
        self.model_config = read_yaml(model_config_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    def get_data_validation_config(self) -> DataValidationConfig:
        config = self.config.data_validation
        schema = self.schema.COLUMNS
        create_directories([config.root_dir])

        return DataValidationConfig(
            root_dir=config.root_dir,
            raw_data_path=config.raw_data_path,
            processed_data_path=config.processed_data_path,
            all_schema=schema,
            STATUS_FILE=config.STATUS_FILE,
        )

import pandas as pd
from src.logger import logger

class DataValidation:
    def __init__(self, config: DataValidationConfig):
        self.config = config

    def preprocess_data(self):
        df = pd.read_csv(self.config.raw_data_path)
        logger.info(f"Data loaded from {self.config.raw_data_path}")

        df["Gender"] = df["Gender"].replace("Fe Male", "Female")
        df["MaritalStatus"] = df["MaritalStatus"].replace("Single", "Unmarried")

        df["Age"].fillna(df["Age"].median(), inplace=True)
        df["TypeofContact"].fillna(df["TypeofContact"].mode()[0], inplace=True)
        df["DurationOfPitch"].fillna(df["DurationOfPitch"].median(), inplace=True)
        df["NumberOfFollowups"].fillna(df["NumberOfFollowups"].mode()[0], inplace=True)
        df["PreferredPropertyStar"].fillna(
            df["PreferredPropertyStar"].mode()[0], inplace=True
        )
        df["NumberOfTrips"].fillna(0, inplace=True)
        df["NumberOfChildrenVisiting"].fillna(
            df["NumberOfChildrenVisiting"].mode()[0], inplace=True
        )
        df["MonthlyIncome"].fillna(df["MonthlyIncome"].median(), inplace=True)

        df.drop("CustomerID", axis=1, inplace=True)
        df["TotalVisiting"] = (
            df["NumberOfChildrenVisiting"] + df["NumberOfPersonVisiting"]
        )
        df.drop(
            ["NumberOfChildrenVisiting", "NumberOfPersonVisiting"], axis=1, inplace=True
        )

        logger.info("Data preprocessed")

        return df

    def save_data(self, df):
        df.to_csv(self.config.processed_data_path, index=False)
        logger.info(f"Data saved at {self.config.processed_data_path}")

    def validate_all_columns(self) -> bool:
        try:
            validation_status = None

            data = pd.read_csv(self.config.processed_data_path)
            all_cols = list(data.columns)

            all_schema = self.config.all_schema.keys()

            for col in all_cols:
                if col not in all_schema:
                    validation_status = False
                    with open(self.config.STATUS_FILE, 'w') as f:
                        f.write(f"Validation status: {validation_status}")
                else:
                    validation_status = True
                    with open(self.config.STATUS_FILE, 'w') as f:
                        f.write(f"Validation status: {validation_status}")

            return validation_status

        except Exception as e:
            raise e


try:
    config = ConfigurationManager()
    data_validation_config = config.get_data_validation_config()
    data_validation = DataValidation(config=data_validation_config)
    df = data_validation.preprocess_data()
    data_validation.save_data(df)
    data_validation.validate_all_columns()
except Exception as e:
    raise e
