==> notebooks/03_data_transformation.ipynb <==
import os

%pwd

os.chdir("../")

%pwd

import numpy as np
import pandas as pd
import warnings

warnings.filterwarnings("ignore")

df = pd.read_csv("./artifacts/data_validation/travel_processed.csv")
df.head()

X = df.drop("ProdTaken", axis=1)
y = df["ProdTaken"]

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.30, random_state=42
)

cat_features = X.select_dtypes(include="object").columns
num_features = X.select_dtypes(exclude="object").columns

from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.compose import ColumnTransformer

numeric_transformer = StandardScaler()
oh_transformer = OneHotEncoder(drop="first")

preprocessor = ColumnTransformer(
    [
        ("OneHotEncoder", oh_transformer, cat_features),
        ("StandardScaler", numeric_transformer, num_features),
    ]
)

cat_features, num_features

df.apply(lambda col: col.unique())

df["Occupation"].value_counts()

X_train = preprocessor.fit_transform(X_train)
X_test = preprocessor.transform(X_test)

X_train

from dataclasses import dataclass
from pathlib import Path


@dataclass(frozen=True)
class DataTransformationConfig:
    root_dir: Path
    processed_data_path: Path
    preprocessor_model_name: str

from src.constants import *
from src.utils.common import read_yaml, create_directories

class ConfigurationManager:
    def __init__(
        self,
        config_filepath=CONFIG_FILE_PATH,
        model_config_filepath=MODEL_CONFIG_FILE_PATH,
        schema_filepath=SCHEMA_FILE_PATH,
    ):

        self.config = read_yaml(config_filepath)
        self.model_config = read_yaml(model_config_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    def get_data_transformation_config(self) -> DataTransformationConfig:
        config = self.config.data_transformation

        create_directories([config.root_dir])

        return DataTransformationConfig(
            root_dir=config.root_dir,
            processed_data_path=config.processed_data_path,
            preprocessor_model_name=config.preprocessor_model_name,
        )

from src.logger import logger
from sklearn.model_selection import train_test_split
import joblib

class DataTransformation:
    def __init__(self, config: DataTransformationConfig):
        self.config = config

    def train_test_spliting(self):
        data = pd.read_csv(self.config.processed_data_path)

        train, test = train_test_split(data, test_size=0.30, random_state=42)

        train.to_csv(os.path.join(self.config.root_dir, "train.csv"), index=False)
        test.to_csv(os.path.join(self.config.root_dir, "test.csv"), index=False)

        logger.info("Splited data into training and test sets")
        logger.info(train.shape)
        logger.info(test.shape)

    def fit_and_save_column_transformer(self):
        train = pd.read_csv(os.path.join(self.config.root_dir, "train.csv"))

        train_X = train.drop("ProdTaken", axis=1)

        cat_features = train_X.select_dtypes(include="object").columns
        num_features = train_X.select_dtypes(exclude="object").columns

        numeric_transformer = StandardScaler()
        oh_transformer = OneHotEncoder(drop="first")

        preprocessor = ColumnTransformer(
            [
                ("OneHotEncoder", oh_transformer, cat_features),
                ("StandardScaler", numeric_transformer, num_features),
            ]
        )

        preprocessor.fit(train_X)

        preprocessor_model_path = os.path.join(
            self.config.root_dir, self.config.preprocessor_model_name
        )
        joblib.dump(preprocessor, preprocessor_model_path)
        logger.info(f"Saved preprocessor model at {preprocessor_model_path}")

try:
    config = ConfigurationManager()
    data_transformation_config = config.get_data_transformation_config()
    data_transformation = DataTransformation(config=data_transformation_config)
    data_transformation.train_test_spliting()
    data_transformation.fit_and_save_column_transformer()
except Exception as e:
    raise e
