==> notebooks/04_model_trainer.ipynb <==
import os

%pwd

os.chdir("../")

%pwd

import numpy as np
import pandas as pd
import warnings

warnings.filterwarnings("ignore")

train = pd.read_csv("./artifacts/data_transformation/train.csv")
test = pd.read_csv("./artifacts/data_transformation/test.csv")
train.shape, test.shape

X_train = train.drop("ProdTaken", axis=1)
y_train = train["ProdTaken"]
X_test = test.drop("ProdTaken", axis=1)
y_test = test["ProdTaken"]

from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.compose import ColumnTransformer
import joblib

numeric_transformer = StandardScaler()
oh_transformer = OneHotEncoder(drop="first")

preprocessor = joblib.load("./artifacts/data_transformation/preprocessor.joblib")

X_train = preprocessor.transform(X_train)
X_test = preprocessor.transform(X_test)

X_train.shape, X_test.shape

from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import AdaBoostClassifier
from xgboost import XGBClassifier
from sklearn.svm import SVC
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import GaussianNB
from sklearn.metrics import (
    accuracy_score,
    precision_score,
    recall_score,
    f1_score,
    roc_auc_score,
)

def compute_metrics(y_train, y_train_pred, y_test, y_test_pred):
    """
    Compute performance metrics for training and test predictions.

    Parameters:
    - y_train (array-like): True labels for training data.
    - y_train_pred (array-like): Predicted labels for training data.
    - y_test (array-like): True labels for test data.
    - y_test_pred (array-like): Predicted labels for test data.

    Returns:
    - metrics (dict): Dictionary containing metrics for both training and test data.
    """
    metrics = {
        "Training Set": {
            "Accuracy": accuracy_score(y_train, y_train_pred),
            "F1 Score": f1_score(y_train, y_train_pred, average="weighted"),
            "Precision": precision_score(y_train, y_train_pred, average="weighted"),
            "Recall": recall_score(y_train, y_train_pred, average="weighted"),
            "Roc Auc Score": roc_auc_score(y_train, y_train_pred),
        },
        "Test Set": {
            "Accuracy": accuracy_score(y_test, y_test_pred),
            "F1 Score": f1_score(y_test, y_test_pred, average="weighted"),
            "Precision": precision_score(y_test, y_test_pred, average="weighted"),
            "Recall": recall_score(y_test, y_test_pred, average="weighted"),
            "Roc Auc Score": roc_auc_score(y_test, y_test_pred),
        },
    }

    return metrics

# Define your models
models = {
    "Logistic Regression": LogisticRegression(),
    "Decision Tree": DecisionTreeClassifier(),
    "Random Forest": RandomForestClassifier(),
    "Gradient Boost": GradientBoostingClassifier(),
    "Adaboost": AdaBoostClassifier(),
    "Xgboost": XGBClassifier(),
    "SVC": SVC(),
    "Gaussian Naive Bayes": GaussianNB(),
}

# Fit and evaluate each model
for model_name, model in models.items():
    print(f"Evaluating model: {model_name}")

    # Fit the model
    model.fit(X_train, y_train)

    # Make predictions
    y_train_pred = model.predict(X_train)
    y_test_pred = model.predict(X_test)

    # Compute metrics
    metrics = compute_metrics(y_train, y_train_pred, y_test, y_test_pred)

    # Print metrics
    print("Model performance for Training set")
    for metric_name, value in metrics["Training Set"].items():
        print(f"- {metric_name}: {value:.4f}")

    print("----------------------------------")

    print("Model performance for Test set")
    for metric_name, value in metrics["Test Set"].items():
        print(f"- {metric_name}: {value:.4f}")

    print("=" * 35)
    print("\n")

random_forest_params = {
    "max_depth": [5, 8, 15, None, 10],
    "max_features": [5, 7, "auto", 8],
    "min_samples_split": [2, 8, 15, 20],
    "n_estimators": [100, 200, 500, 1000],
}
xgboost_params = {
    "learning_rate": [0.1, 0.01],
    "max_depth": [5, 8, 12, 20, 30],
    "n_estimators": [100, 200, 300],
    "colsample_bytree": [0.5, 0.8, 1, 0.3, 0.4],
}
decision_tree_params = {
    "criterion": ["gini", "entropy", "log_loss"],
    "splitter": ["best", "random"],
    "max_depth": [1, 2, 3, 4, 5],
    "max_features": ["auto", "sqrt", "log2"],
}

from sklearn.model_selection import RandomizedSearchCV

randomcv_models = [
    ("Random Forest", RandomForestClassifier(), random_forest_params),
    ("Xgboost", XGBClassifier(), xgboost_params),
    ("Decision Tree", DecisionTreeClassifier(), decision_tree_params),
]

model_param = {}
for name, model, params in randomcv_models:
    random = RandomizedSearchCV(
        estimator=model,
        param_distributions=params,
        n_iter=100,
        cv=3,
        verbose=2,
        n_jobs=-1,
    )
    random.fit(X_train, y_train)
    model_param[name] = random.best_params_

for model_name in model_param:
    print(f"---------------- Best Params for {model_name} -------------------")
    print(model_param[model_name])

models = {
    "Random Forest": RandomForestClassifier(
        n_estimators=1000, min_samples_split=2, max_features=8, max_depth=15
    ),
    "Xgboost": XGBClassifier(
        n_estimators=300, max_depth=8, learning_rate=0.1, colsample_bytree=0.5
    ),
    "Decision Tree": DecisionTreeClassifier(
        splitter="random", max_features="sqrt", max_depth=4, criterion="log_loss"
    ),
}

for model_name, model in models.items():
    print(f"Evaluating model: {model_name}")

    # Fit the model
    model.fit(X_train, y_train)

    # Make predictions
    y_train_pred = model.predict(X_train)
    y_test_pred = model.predict(X_test)

    # Compute metrics
    metrics = compute_metrics(y_train, y_train_pred, y_test, y_test_pred)

    # Print metrics
    print("Model performance for Training set")
    for metric_name, value in metrics["Training Set"].items():
        print(f"- {metric_name}: {value:.4f}")

    print("----------------------------------")

    print("Model performance for Test set")
    for metric_name, value in metrics["Test Set"].items():
        print(f"- {metric_name}: {value:.4f}")

    print("=" * 35)
    print("\n")

xgboost_classfier = XGBClassifier(
    n_estimators=300, max_depth=8, learning_rate=0.1, colsample_bytree=0.5
)
xgboost_classfier.fit(X_train, y_train)

from dataclasses import dataclass
from pathlib import Path


@dataclass(frozen=True)
class ModelTrainerConfig:
    root_dir: Path
    train_data_path: Path
    test_data_path: Path
    trained_model_name: str
    preprocessor_model_path: Path
    model_name: str
    model_params: dict
    target_column: str

from src.constants import *
from src.constants import MODEL_CONFIG_FILE_PATH
from src.utils.common import read_yaml, create_directories


yaml_path = Path("./model_config.yaml")
yaml_data = read_yaml(yaml_path)

print(yaml_data)

class ConfigurationManager:
    def __init__(
        self,
        config_filepath=CONFIG_FILE_PATH,
        model_config_filepath=MODEL_CONFIG_FILE_PATH,
        schema_filepath=SCHEMA_FILE_PATH,
    ):

        self.config = read_yaml(config_filepath)
        self.model_config = read_yaml(model_config_filepath)
        self.schema = read_yaml(schema_filepath)

        create_directories([self.config.artifacts_root])

    def get_model_trainer_config(self) -> ModelTrainerConfig:
        config = self.config.model_trainer
        model_name = self.model_config.model
        params = self.model_config.params
        schema = self.schema.TARGET_COLUMN
        
        create_directories([config.root_dir])

        model_trainer_config = ModelTrainerConfig(
            root_dir=config.root_dir,
            train_data_path=config.train_data_path,
            test_data_path=config.test_data_path,
            trained_model_name=config.trained_model_name,
            preprocessor_model_path = config.preprocessor_model_path,
            model_name=model_name,
            model_params=params,
            target_column=schema.name,
        )

        return model_trainer_config

from src.logger import logger
import joblib

class ModelTrainer:
    def __init__(self, config: ModelTrainerConfig):
        self.config = config

    def train(self):
        models = {
            "Logistic Regression": LogisticRegression,
            "Decision Tree": DecisionTreeClassifier,
            "Random Forest": RandomForestClassifier,
            "Gradient Boost": GradientBoostingClassifier,
            "Adaboost": AdaBoostClassifier,
            "Xgboost": XGBClassifier,
            "SVC": SVC,
            "Gaussian Naive Bayes": GaussianNB,
        }

        # Load training data
        train_data = pd.read_csv(self.config.train_data_path)
        X_train = train_data.drop([self.config.target_column], axis=1)
        y_train = train_data[[self.config.target_column]]

        # Load preprocessor and transform data
        preprocessor = joblib.load(self.config.preprocessor_model_path)
        X_train = preprocessor.transform(X_train)

        # Load model parameters and initialize model
        model_name = self.config.model_name
        if model_name not in models:
            raise ValueError(f"Model '{model_name}' is not supported.")

        model_class = models[model_name]
        model_params = self.config.model_params

        # Instantiate the model with parameters
        prod_model = model_class(**model_params)

        # Fit the model
        prod_model.fit(X_train, y_train)

        # Save the trained model
        joblib.dump(
            prod_model,
            os.path.join(self.config.root_dir, self.config.trained_model_name),
        )

try:
    config = ConfigurationManager()
    model_trainer_config = config.get_model_trainer_config()
    model_trainer_config = ModelTrainer(config=model_trainer_config)
    model_trainer_config.train()
except Exception as e:
    raise e
